+++ rodbus/src/common/mod.rs
@@ -6,5 +6,6 @@ pub(crate) mod buffer;
 pub(crate) mod cursor;
 pub(crate) mod frame;
 mod parse;
+pub(crate) mod pdu;
 pub(crate) mod phys;
 mod serialize;
