+++ rodbus/src/server/response.rs
@@ -39,34 +39,6 @@ where
     }
 }
 
-pub(crate) struct Response<'a, T>
-where
-    T: Serialize,
-{
-    function: FunctionCode,
-    body: &'a T,
-}
-
-impl<'a, T> Response<'a, T>
-where
-    T: Serialize,
-{
-    pub(crate) fn new(function: FunctionCode, body: &'a T) -> Self {
-        Response { function, body }
-    }
-}
-
-impl<'a, T> Serialize for Response<'a, T>
-where
-    T: Serialize,
-{
-    fn serialize(&self, cursor: &mut WriteCursor) -> Result<(), RequestError> {
-        cursor.write_u8(self.function.get_value())?;
-        self.body.serialize(cursor)?;
-        Ok(())
-    }
-}
-
 pub(crate) struct ErrorResponse {
     pub(crate) function: u8,
     pub(crate) exception: ExceptionCode,
